bug-fix

+++ src/core/user/client/clientController/deleteClientById.js
@@ -2,8 +2,9 @@ const createError = require('http-errors');
 const clientService = require("../clientService");
 
 const deleteClient = async (req, res) => {
-    const { _id: id } = req.user;
-  
+  if (!req.user) {
+    return res.status(401).json({ message: 'Unauthorized' });
+  }
     const user = await clientService.deleteClientById(id);
     if (!user) {
       throw createError(500, "Server error");
